requirements.txt
PyMuPDF>=1.25.0
Pillow>=11.0.0
pytesseract>=0.3.13
openai>=1.68.0
PyYAML>=6.0.0
fastapi
uvicorn
python-multipart
langgraph